requirements.txt
Django==2.1.7
Pillow==6.1.0
pytz==2019.1
